Rename files to stick with naming convetion
